+++ erudit/fedora/views/generic.py
@@ -7,7 +7,6 @@ that involve Fedora and datastreams.
 
 import logging
 
-from django.core.cache import caches
 from django.core.exceptions import ImproperlyConfigured
 from django.http import Http404
 from django.http import HttpResponse
@@ -16,8 +15,8 @@ from django.views.generic.detail import SingleObjectMixin
 from eulfedora.util import RequestFailed
 from requests.exceptions import ConnectionError
 
-from ...conf import settings as erudit_settings
 from ..repository import api
+from ..shortcuts import get_cached_datastream_content
 
 logger = logging.getLogger(__name__)
 
@@ -27,7 +26,6 @@ class FedoraFileDatastreamView(SingleObjectMixin, View):
     The FedoraFileDatastreamView CBV can be used to expose Fedora file datastreams
     through Django views.
     """
-    cache_timeout = 60 * 60  # This is expressed in seconds
     http_method_names = ['get', ]
 
     # The following attributes should be specified on each subclass
@@ -80,12 +78,11 @@ class FedoraFileDatastreamView(SingleObjectMixin, View):
         Writes the content of the fedora object's datastream to an HttpResponse object
         and return it.
         """
-        cache = self.get_cache()
-        content_key = 'erudit-fedora-file-{pid}'.format(pid=fedora_object.pid)
-        content = cache.get(content_key)
-
+        if self.datastream_name is None:
+            raise ImproperlyConfigured(
+                '{cls} is missing a datastream. Define {cls}.datastream_name '
+                'or override {cls}.get_datastream_content().'.format(cls=self.__class__.__name__))
         try:
-            assert content is None
             content = self.get_datastream_content(fedora_object)
         except RequestFailed:  # pragma: no cover
             # This means that the datastream content could not be retrieved
@@ -99,12 +96,6 @@ class FedoraFileDatastreamView(SingleObjectMixin, View):
                 'request': self.request,
             })
             raise Http404
-        except AssertionError:
-            # We've just retrieved the content of the file from the file-based cache!
-            pass
-        else:
-            # Puts the content of the file in the file-based cache!
-            cache.set(content_key, content, self.cache_timeout)
 
         response = self.get_response_object(fedora_object)
         self.write_datastream_content(response, content)
@@ -135,10 +126,6 @@ class FedoraFileDatastreamView(SingleObjectMixin, View):
                 'Define {cls}.content_type.'.format(cls=self.__class__.__name__))
         return self.content_type
 
-    def get_cache(self):
-        """ Returns the cache to use to store file contents. """
-        return caches[erudit_settings.FEDORA_FILEBASED_CACHE_NAME]
-
     def get_datastream_content(self, fedora_object):
         """
         Returns the content of the considered Fedora datastream.
@@ -146,13 +133,9 @@ class FedoraFileDatastreamView(SingleObjectMixin, View):
         By default this requires `self.datastream_name` to be specified but
         subclasses can override this to change this.
         """
-        if self.datastream_name is None:
-            raise ImproperlyConfigured(
-                '{cls} is missing a datastream. Define {cls}.datastream_name '
-                'or override {cls}.get_datastream_content().'.format(cls=self.__class__.__name__))
         try:
             assert hasattr(fedora_object, self.datastream_name)
-            return getattr(fedora_object, self.datastream_name).content
+            return get_cached_datastream_content(fedora_object, self.datastream_name)
         except (AssertionError, AttributeError):
             raise ImproperlyConfigured(
                 'The `{ds}` datastream cannot be retrieved on the {fedoracls} digital object. '